make rate-limiting query manager
Our poll loop now spams calls to query_man way faster than we actually
want to query the chain, but query_man blocks most of these. This is to
simulate the situation where request handlers I write all want to query
the same information, which might not have been mined yet. This might
not be needed in practice, but only because I will likely use
subscriptions instead, since it really is a push type of event rather
than a pull type of event, but either way, this module will also be the
place to handle subscriptions, so this is a good test of the same
architecture.