ground truth module
This loads and formats the chain state, while keeping the work in a
different module. This will completely re-synchronize the state every
minute, so that any backdoor use of the contracts will still make its
way to the front end. This also gives me a centralised place to build
poll state if a new poll gets registered in the frontend.